refactor(canon): split virtual_ts root into cohesive modules
